Restructure vocab into module.
Move subword and simple vocabulary implementations to different
files.
